agent-development-kit-brndn/requirements.txt
google-adk[database]
yfinance
psutil
litellm
google-generativeai
python-dotenv